move cron to service and controller
